+++ RaidLib/Simulator/ClanBossInBattle.cs
@@ -41,9 +41,9 @@ namespace RaidLib.Simulator
 
         }
 
-        public float TurnMeter { get; private set; }
+        public double TurnMeter { get; private set; }
 
-        public float TurnMeterIncreaseOnClockTick { get; private set; }
+        public double TurnMeterIncreaseOnClockTick { get; private set; }
 
         public Skill TakeTurn()
         {
